HearingsHelperTest
