+++ src/test/java/uk/gov/hmcts/ethos/replacement/docmosis/helpers/HelperTest.java
@@ -96,81 +96,7 @@ public class HelperTest {
         assertEquals(activeRespondentsFound, activeRespondents.size());
     }
 
-    @Test
-    public void hearingMidEventValidationNumberError() {
 
-        caseDetails1.getCaseData().getHearingCollection().get(0).getValue().setHearingNumber(null);
-
-        assertEquals(1, HearingsHelper.hearingMidEventValidation(caseDetails1.getCaseData()).size());
-
-        assertEquals(HEARING_CREATION_NUMBER_ERROR,
-                HearingsHelper.hearingMidEventValidation(caseDetails1.getCaseData()).get(0));
-
-        caseDetails1.getCaseData().getHearingCollection().get(0).getValue().setHearingNumber("");
-
-        assertEquals(1, HearingsHelper.hearingMidEventValidation(caseDetails1.getCaseData()).size());
-
-        assertEquals(HEARING_CREATION_NUMBER_ERROR,
-                HearingsHelper.hearingMidEventValidation(caseDetails1.getCaseData()).get(0));
-
-    }
-
-    @Test
-    public void hearingMidEventValidationDayError() {
-
-        caseDetails1.getCaseData().getHearingCollection().get(0).getValue()
-                .getHearingDateCollection().get(0).getValue().setListedDate(null);
-
-        assertEquals(1, HearingsHelper.hearingMidEventValidation(caseDetails1.getCaseData()).size());
-
-        assertEquals(HEARING_CREATION_DAY_ERROR, HearingsHelper.hearingMidEventValidation(caseDetails1.getCaseData()).get(0));
-
-        caseDetails1.getCaseData().getHearingCollection().get(0).getValue()
-                .getHearingDateCollection().get(0).getValue().setListedDate("");
-
-        assertEquals(1, HearingsHelper.hearingMidEventValidation(caseDetails1.getCaseData()).size());
-
-        assertEquals(HEARING_CREATION_DAY_ERROR, HearingsHelper.hearingMidEventValidation(caseDetails1.getCaseData()).get(0));
-
-        caseDetails1.getCaseData().getHearingCollection().get(0).getValue()
-                .setHearingDateCollection(null);
-
-        assertEquals(0, HearingsHelper.hearingMidEventValidation(caseDetails1.getCaseData()).size());
-
-    }
-
-    @Test
-    public void updatePostponedDate() {
-
-        caseDetails1.getCaseData().getHearingCollection().get(0).getValue()
-                .getHearingDateCollection().get(0).getValue().setHearingStatus(HEARING_STATUS_POSTPONED);
-
-        HearingsHelper.updatePostponedDate(caseDetails1.getCaseData());
-
-        assertNotNull(caseDetails1.getCaseData().getHearingCollection().get(0).getValue()
-                .getHearingDateCollection().get(0).getValue().getPostponedDate());
-
-        caseDetails1.getCaseData().getHearingCollection().get(0).getValue()
-                .getHearingDateCollection().get(0).getValue().setHearingStatus(HEARING_STATUS_SETTLED);
-
-        HearingsHelper.updatePostponedDate(caseDetails1.getCaseData());
-
-        assertNull(caseDetails1.getCaseData().getHearingCollection().get(0).getValue()
-                .getHearingDateCollection().get(0).getValue().getPostponedDate());
-    }
-
-    @Test
-    public void findDateOfHearingTest() {
-        var hearingDate = caseDetails1.getCaseData().getHearingCollection().get(0).getValue()
-                .getHearingDateCollection().get(0).getValue().getListedDate().substring(0, 10);
-        var hearingNumber = findHearingNumber(caseDetails1.getCaseData(), hearingDate);
-        assertEquals(hearingNumber, caseDetails1.getCaseData().getHearingCollection().get(0).getValue().getHearingNumber());
-    }
-
-    @Test
-    public void findDateOfHearing_DateNotInHearing() {
-        assertNull(findHearingNumber(caseDetails1.getCaseData(), "1970-10-01"));
-    }
